Python/Speec-to-Text-WebApp/requirements.txt
flask
openai-whisper==20230314
